Fixed small bugs and added small corrections

+++ src/FusionEKF.h
@@ -38,8 +38,8 @@ private:
   // previous timestamp
   long long previous_timestamp_;
 
-  int noise_ax;
-  int noise_ay;
+  float noise_ax;
+  float noise_ay;
 
   // tool object used to compute Jacobian and RMSE
   Tools tools;
